requirements.txt
fastapi
uvicorn[standard]
pydantic

httpx
aiohttp

websockets
python-socketio

psycopg2-binary
asyncpg
sqlalchemy
alembic

redis

geopy
shapely
pyproj

python-dateutil
pytz

loguru

python-multipart
python-dotenv
jinja2
msgpack
jupyter
ipykernel

boto3

streamlit
streamlit-folium
folium
pandas
